feat(config): Add use_gui flag to select execution mode (#8)
Adds a `use_gui` boolean flag to `base_config.json` to allow switching between GUI and headless execution from a single entry point.

- The `run_study.py` script now reads the `use_gui` flag.
- If true, it launches the standard GUI.
- If false, it runs in headless mode, using a console logger for output.
- The logic from the deprecated `run_study_no_gui.py` has been merged into `run_study.py`.
- The `run_study_no_gui.py` file has been removed to consolidate the entry points.

Closes #8
